requirements.txt
asgiref==3.7.2
attrs==24.2.0
black==24.8.0
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-vite==3.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.2
psycopg==3.1.18
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
